+++ src/planning_toolbox/sketchup/plugin/planning_toolbox_sketchup/mcp_bridge.rb
@@ -16,7 +16,7 @@ module PlanningToolbox
   module SketchUpMcpBridge
     HOST = '127.0.0.1'
     DEFAULT_PORT = 8765
-    VERSION = '0.61.0'
+    VERSION = '0.61.1'
     WATCHDOG_INTERVAL = 3.0
     CONFIG_DIR = File.join(ENV.fetch('APPDATA', Dir.tmpdir), 'PlanningToolbox')
     CONFIG_PATH = ENV.fetch(